requirements.txt
langchain-openai
langchain-community==0.3.14
python-dotenv==1.0.1
tiktoken==0.8.0
faiss-cpu==1.9.0.post1
openai==1.59.7
